chore(databases-collections): update drop database and collection modal components to LeafyGreen COMPASS-6223 (#3729)
