+++ packages/databases-collections/package.json
@@ -51,7 +51,6 @@
     "@mongodb-js/compass-logging": "^1.1.1",
     "@mongodb-js/databases-collections-list": "^1.3.0",
     "compass-preferences-model": "^2.3.0",
-    "hadron-react-components": "^6.3.0",
     "react": "^16.14.0"
   },
   "devDependencies": {
@@ -72,6 +71,8 @@
     "hadron-ipc": "^3.1.0",
     "lodash": "^4.17.15",
     "mocha": "^8.4.0",
+    "mongodb": "^4.10.0",
+    "mongodb-collection-model": "^5.2.0",
     "mongodb-data-service": "^22.2.0",
     "mongodb-instance-model": "^12.2.0",
     "mongodb-ns": "^2.4.0",
@@ -92,7 +93,6 @@
     "@mongodb-js/compass-logging": "^1.1.1",
     "@mongodb-js/databases-collections-list": "^1.3.0",
     "compass-preferences-model": "^2.3.0",
-    "hadron-react-components": "^6.3.0",
     "react": "^16.14.0"
   }
 }
