+++ packages/databases-collections/src/components/create-database-modal/create-database-modal.jsx
@@ -107,7 +107,6 @@ class CreateDatabaseModal extends PureComponent {
         )}
         trackingId="create_database_modal"
         data-testid="create-database-modal"
-
       >
         <CollectionFields
           serverVersion={this.props.serverVersion}
