--- packages/databases-collections/src/components/drop-collection-modal/index.js
@@ -1,3 +1,0 @@
-import MappedDropCollectionModal, { DropCollectionModal } from './drop-collection-modal';
-export default MappedDropCollectionModal;
-export { DropCollectionModal };
