+++ packages/databases-collections/src/components/drop-collection-plugin.tsx
@@ -1,0 +1,15 @@
+import React from 'react';
+import { Provider } from 'react-redux';
+
+import DropCollectionModal from './drop-collection-modal';
+import store from '../stores/drop-collection';
+
+function DropCollectionPlugin() {
+  return (
+    <Provider store={store}>
+      <DropCollectionModal />
+    </Provider>
+  );
+}
+
+export default DropCollectionPlugin;
