--- packages/databases-collections/src/components/drop-collection-plugin/drop-collection-plugin.jsx
@@ -1,23 +1,0 @@
-import React, { Component } from 'react';
-import { Provider } from 'react-redux';
-import DropCollectionModal from '../drop-collection-modal';
-import store from '../../stores/drop-collection';
-
-class DropCollectionPlugin extends Component {
-  static displayName = 'DropCollectionPlugin';
-
-  /**
-   * Connect the Plugin to the store and render.
-   *
-   * @returns {React.Component} The rendered component.
-   */
-  render() {
-    return (
-      <Provider store={store}>
-        <DropCollectionModal />
-      </Provider>
-    );
-  }
-}
-
-export default DropCollectionPlugin;
