--- packages/databases-collections/src/components/drop-collection-plugin/index.js
@@ -1,2 +1,0 @@
-import DropCollectionPlugin from './drop-collection-plugin';
-export default DropCollectionPlugin;
