--- packages/databases-collections/src/components/drop-database-modal/drop-database-modal.module.less
@@ -1,17 +1,0 @@
-.drop-database-modal {
-  &-confirm {
-    i {
-      margin: 0px 4px 0px 2px;
-    }
-
-    &-namespace {
-      font-weight: bold;
-      margin: 0px 4px 0px 4px;
-    }
-
-    &-name {
-      font-weight: bold;
-      margin-left: 4px;
-    }
-  }
-}
