--- packages/databases-collections/src/components/drop-database-modal/index.js
@@ -1,3 +1,0 @@
-import MappedDropDatabaseModal, { DropDatabaseModal } from './drop-database-modal';
-export default MappedDropDatabaseModal;
-export { DropDatabaseModal };
