--- packages/databases-collections/src/components/drop-database-plugin/index.js
@@ -1,2 +1,0 @@
-import DropDatabasePlugin from './drop-database-plugin';
-export default DropDatabasePlugin;
