+++ packages/databases-collections/src/modules/drop-collection/name.ts
@@ -1,0 +1,7 @@
+type State = string;
+
+export const INITIAL_STATE: State = '';
+
+export default function reducer(state = INITIAL_STATE): State {
+  return state;
+}
