--- packages/databases-collections/src/modules/index.js
@@ -1,7 +1,0 @@
-import collectionsReducer from './collections';
-import databasesReducer from './databases';
-
-export {
-  collectionsReducer,
-  databasesReducer
-};
