+++ packages/databases-collections/src/modules/reset.ts
@@ -1,0 +1,14 @@
+/**
+ * The reset action name.
+ */
+export const RESET = 'databases-collections/reset';
+
+/**
+ * Reset the state action.
+ *
+ * @return {Object} The action creator.
+ */
+export const reset = () => ({
+  type: RESET
+});
+
