+++ packages/databases-collections/src/stores/drop-collection.js
@@ -1,7 +1,7 @@
 import { createStore, applyMiddleware } from 'redux';
 import thunk from 'redux-thunk';
 import { dataServiceConnected } from '../modules/data-service';
-import reducer, { open } from '../modules/drop-collection';
+import reducer, { open } from '../modules/drop-collection/drop-collection';
 
 const store = createStore(reducer, applyMiddleware(thunk));
